services/auto-remediator/requirements.txt
kafka-python
PyYAML
